+++ changelog.md
@@ -4,6 +4,7 @@
 
 ## v4.18.0 (January 29, 2026)
 - Add max tokens to main settings popup, serving as an override for chat completion tokens.
+- Add order settings to arcs.
 
 ## v4.17.0 (January 27, 2026)
 - Add ability to manually set highest processed message for chats with `/stmb-set-highest x`.